Modified the shooter pneumatics subsystem to use the MomentaryDoubleSolenoid.
If the air pressure drops too low, and a command is issued to lower the
hood, it won't actually dump the air out of the actuator. Subsequent
orders issued will not cause the actuator to move either. == hood won't
ever lower no matter how many times you hit the button.

MomentaryDoubleSolenoid will allow the commands to be re-issued as the
relay outputs are periodically de-energized.